Basic implementation added.
TODO:
1. rewrite comparator to return multiple equivalent results;
2. rewrite average double calculation to BigDecimal
3. add Java Docs
4. add Unit tests

+++ src/main/java/by/vk/betting/api/analytic/dto/result/ResultHolder.java
@@ -1,4 +1,5 @@
 package by.vk.betting.api.analytic.dto.result;
 
-public record ResultHolder(String team, int amount) {
+
+public record ResultHolder(String team, double amount) {
 }
